refactor: move common into scoped directory
