+++ src/components/common/Hero.tsx
