+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Layout from '../components/layout';
-import SEO from '../components/seo';
+import Layout from '@/components/common/Layout';
+import SEO from '@/components/common/SEO';
 
 const NotFoundPage = () => (
   <Layout>
